backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
curlify==2.2.1
facebook-business==16.0.0
fastapi==0.92.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.5
pycountry==22.3.5
pydantic==1.10.5
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
yarl==1.8.2
